--- alembic/versions/07739cb4b544_useraccessdetails.py
@@ -1,39 +1,0 @@
-"""UserAccessDetails
-
-Revision ID: 07739cb4b544
-Revises: 1ccf26e56a93
-Create Date: 2021-04-06 13:22:36.473998
-
-"""
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import mysql
-
-# revision identifiers, used by Alembic.
-revision = '07739cb4b544'
-down_revision = '1ccf26e56a93'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('user_access_details',
-    sa.Column('row_id', sa.BIGINT(), autoincrement=True, nullable=False),
-    sa.Column('row_created', mysql.TIMESTAMP(), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
-    sa.Column('row_updated', mysql.TIMESTAMP(), server_default=sa.text('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'), nullable=False),
-    sa.Column('wallet_address', sa.VARCHAR(length=50), nullable=False),
-    sa.Column('email', sa.VARCHAR(length=120), nullable=False),
-    sa.Column('name', sa.VARCHAR(length=60), nullable=False),
-    sa.Column('login_time', mysql.TIMESTAMP(), nullable=False),
-    sa.PrimaryKeyConstraint('row_id')
-    )
-    op.create_index(op.f('ix_user_access_details_wallet_address'), 'user_access_details', ['wallet_address'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_user_access_details_wallet_address'), table_name='user_access_details')
-    op.drop_table('user_access_details')
-    # ### end Alembic commands ###
