--- alembic/versions/1ccf26e56a93_snapshots.py
@@ -1,39 +1,0 @@
-"""Snapshots
-
-Revision ID: 1ccf26e56a93
-Revises: 
-Create Date: 2021-04-06 13:20:37.237352
-
-"""
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import mysql
-
-# revision identifiers, used by Alembic.
-revision = '1ccf26e56a93'
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('snapshots',
-    sa.Column('row_id', sa.BIGINT(), autoincrement=True, nullable=False),
-    sa.Column('row_created', mysql.TIMESTAMP(), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
-    sa.Column('row_updated', mysql.TIMESTAMP(), server_default=sa.text('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'), nullable=False),
-    sa.Column('address', sa.VARCHAR(length=50), nullable=False),
-    sa.Column('block_number', sa.BIGINT(), nullable=False),
-    sa.Column('balance_in_cogs', sa.BIGINT(), nullable=False),
-    sa.Column('snapshot_date', sa.Date(), nullable=False),
-    sa.PrimaryKeyConstraint('row_id')
-    )
-    op.create_index(op.f('ix_snapshots_address'), 'snapshots', ['address'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_snapshots_address'), table_name='snapshots')
-    op.drop_table('snapshots')
-    # ### end Alembic commands ###
