--- alembic/versions/c84b6c2f1332_transferinfo.py
@@ -1,42 +1,0 @@
-"""TransferInfo
-
-Revision ID: c84b6c2f1332
-Revises: e10c9a0ac902
-Create Date: 2021-04-06 13:24:26.923964
-
-"""
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import mysql
-
-# revision identifiers, used by Alembic.
-revision = 'c84b6c2f1332'
-down_revision = 'e10c9a0ac902'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('transfer_info',
-    sa.Column('row_id', sa.BIGINT(), autoincrement=True, nullable=False),
-    sa.Column('row_created', mysql.TIMESTAMP(), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
-    sa.Column('row_updated', mysql.TIMESTAMP(), server_default=sa.text('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'), nullable=False),
-    sa.Column('wallet_address', sa.VARCHAR(length=50), nullable=False),
-    sa.Column('transfer_time', mysql.TIMESTAMP(), nullable=False),
-    sa.Column('transfer_fees', sa.BIGINT(), nullable=False),
-    sa.Column('transfer_transaction', sa.VARCHAR(length=255), nullable=False),
-    sa.Column('transfer_amount_in_cogs', sa.BIGINT(), nullable=False),
-    sa.Column('transfer_status', sa.VARCHAR(length=50), nullable=False),
-    sa.Column('address_type', sa.INTEGER(), nullable=False),
-    sa.PrimaryKeyConstraint('row_id')
-    )
-    op.create_index(op.f('ix_transfer_info_wallet_address'), 'transfer_info', ['wallet_address'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_transfer_info_wallet_address'), table_name='transfer_info')
-    op.drop_table('transfer_info')
-    # ### end Alembic commands ###
