--- alembic/versions/e10c9a0ac902_usercomments.py
@@ -1,37 +1,0 @@
-"""UserComments
-
-Revision ID: e10c9a0ac902
-Revises: 07739cb4b544
-Create Date: 2021-04-06 13:23:20.115033
-
-"""
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import mysql
-
-# revision identifiers, used by Alembic.
-revision = 'e10c9a0ac902'
-down_revision = '07739cb4b544'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('user_comments',
-    sa.Column('row_id', sa.BIGINT(), autoincrement=True, nullable=False),
-    sa.Column('row_created', mysql.TIMESTAMP(), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
-    sa.Column('row_updated', mysql.TIMESTAMP(), server_default=sa.text('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'), nullable=False),
-    sa.Column('wallet_address', sa.VARCHAR(length=50), nullable=False),
-    sa.Column('comment', sa.TEXT(), nullable=False),
-    sa.PrimaryKeyConstraint('row_id')
-    )
-    op.create_index(op.f('ix_user_comments_wallet_address'), 'user_comments', ['wallet_address'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_user_comments_wallet_address'), table_name='user_comments')
-    op.drop_table('user_comments')
-    # ### end Alembic commands ###
